requirements.txt
pytesseract
Pillow
requests
redis
twython
